remove router folder from controller,rename course.js to getCatcourses.js,edit test
